refactor(tmpfs): move `DefaultNode` to tmpfs
